Merge of db-refactoring-dev-bridge into dev

+++ Exebite.DataAccess/Repositories/OrderRepository/IOrderCommandRepository.cs
@@ -1,6 +1,6 @@
 ﻿namespace Exebite.DataAccess.Repositories
 {
-    public interface IOrderCommandRepository : IDatabaseCommandRepository<int, OrderInsertModel, OrderUpdateModel>
+    public interface IOrderCommandRepository : IDatabaseCommandRepository<long, OrderInsertModel, OrderUpdateModel>
     {
     }
 }
