+++ Exebite.DataAccess/Repositories/OrderRepository/IOrderQueryRepository.cs
@@ -6,6 +6,6 @@ namespace Exebite.DataAccess.Repositories
 {
     public interface IOrderQueryRepository : IDatabaseQueryRepository<Order, OrderQueryModel>
     {
-        Either<Error, PagingResult<Order>> GetAllOrdersForRestaurant(int restaruntId, int page, int size);
+        Either<Error, PagingResult<Order>> GetAllOrdersForRestaurant(long restaruntId, int page, int size);
     }
 }
