+++ Exebite.DtoModels/Order/OrderQueryDto.cs
@@ -4,11 +4,10 @@ namespace Exebite.DtoModels
 {
     public class OrderQueryDto : QueryBaseDto
     {
-        public int? Id { get; set; }
+        public long? Id { get; set; }
 
-        // TODO: change this to be our class with data only
         public DateTime? Date { get; set; }
 
-        public int? CustomerId { get; set; }
+        public long? CustomerId { get; set; }
     }
 }
